+++ hacks/glx/companion.c
@@ -391,7 +391,7 @@ cube_handle_event (ModeInfo *mi, XEvent *event)
             event->xbutton.button == Button6 ||
             event->xbutton.button == Button7))
     {
-      gltrackball_mousewheel (bp->trackball, event->xbutton.button, 3,
+      gltrackball_mousewheel (bp->trackball, (int)event->xbutton.button, 3,
                               !event->xbutton.state);
       return True;
     }
@@ -474,7 +474,7 @@ init_cube (ModeInfo *mi)
   glEndList ();
 
 
-  bp->nfloaters = MI_COUNT (mi);
+  bp->nfloaters = (int)(MI_COUNT (mi));
   bp->floaters = (floater *) calloc (bp->nfloaters, sizeof (floater));
 
   for (i = 0; i < bp->nfloaters; i++)
